+++ apps/backoffice-v2/src/common/components/molecules/ProcessTracker/trackers/collection-flow/CollectionFlowTracker.tsx
@@ -6,8 +6,8 @@ import { useCollectionFlowSteps } from './hooks/useCollectionFlowSteps';
 import { useCollectionFlowTrackerItems } from './hooks/useCollectionFlowTrackerItems';
 
 export const CollectionFlowTracker = ({ workflow, plugins, processes }: TTrackerComponentProps) => {
-  const steps = useCollectionFlowSteps({ workflowContext: workflow.context });
-  const items = useCollectionFlowTrackerItems({ steps, workflow });
+  const { steps, isLoading: isLoadingSteps } = useCollectionFlowSteps({ workflowId: workflow.id });
+  const items = useCollectionFlowTrackerItems({ steps, workflow, isLoadingSteps });
 
   return (
     <Tracker workflow={workflow} plugins={plugins} processes={processes}>