+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ballerine/ui
 
+## 0.7.143
+
+### Patch Changes
+
+- bump
+
+## 0.7.142
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.96
+
+## 0.7.141
+
+### Patch Changes
+
+- Bump
+
 ## 0.7.139
 
 ### Patch Changes