+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-ui-sdk
 
+## 1.5.97
+
+### Patch Changes
+
+- Updated dependencies [fed1494bf4a7dde81915d4593765e7d89104c164]
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.96
+
 ## 1.5.96
 
 ### Patch Changes